requirements.txt
Pillow
matplotlib
numpy
scikit-learn
pandas
cvxopt
